Move test around to close its target
