+++ src/routes/github/webhook/webhook-receiver-post.test.ts
@@ -9,11 +9,15 @@ import { deleteRepository } from "~/src/github/repository";
 import { workflowWebhookHandler } from "~/src/github/workflow";
 import { deploymentWebhookHandler } from "~/src/github/deployment";
 import { codeScanningAlertWebhookHandler } from "~/src/github/code-scanning-alert";
+import { GITHUB_CLOUD_HOSTNAME, GITHUB_CLOUD_API_BASEURL } from "~/src/util/get-github-client-config";
+import { envVars } from "config/env";
 
 jest.mock("~/src/middleware/github-webhook-middleware");
 
-const uuid = "97da6b0e-ec61-11ec-8ea0-0242ac120002";
-const webhookSeret = "webhookSecret";
+const EXIST_GHES_UUID = "97da6b0e-ec61-11ec-8ea0-0242ac120002";
+const NON_EXIST_GHES_UUID = "97da6b0e-ec61-11ec-8ea0-0242ac120003";
+const GHES_WEBHOOK_SECRET = "webhookSecret";
+const CLOUD_WEBHOOK_SECRET = envVars.WEBHOOK_SECRET;
 
 describe("webhook-receiver-post", () => {
 
@@ -21,14 +25,26 @@ describe("webhook-receiver-post", () => {
 	let res;
 	let gitHubApp: GitHubServerApp;
 
-	const gitHubAppConfig = () => {
-		return expect.objectContaining({
+	const gitHubAppConfigForCloud= () => {
+		return {
+			gitHubAppId: undefined,
+			appId: parseInt(envVars.APP_ID),
+			clientId: envVars.GITHUB_CLIENT_ID,
+			gitHubBaseUrl: GITHUB_CLOUD_HOSTNAME,
+			gitHubApiUrl: GITHUB_CLOUD_API_BASEURL,
+			uuid: undefined
+		};
+	};
+
+	const gitHubAppConfigForES= () => {
+		return {
 			gitHubAppId: gitHubApp.id,
 			appId: gitHubApp.appId,
 			clientId: gitHubApp.gitHubClientId,
 			gitHubBaseUrl: gitHubApp.gitHubBaseUrl,
+			gitHubApiUrl: gitHubApp.gitHubBaseUrl,
 			uuid: gitHubApp.uuid
-		});
+		};
 	};
 
 	beforeEach(async () => {
@@ -40,13 +56,13 @@ describe("webhook-receiver-post", () => {
 		};
 
 		const payload = {
-			uuid,
+			uuid: EXIST_GHES_UUID,
 			appId: 123,
 			gitHubAppName: "My GitHub Server App",
 			gitHubBaseUrl: "http://myinternalserver.com",
 			gitHubClientId: "lvl.1234",
 			gitHubClientSecret: "myghsecret",
-			webhookSecret: webhookSeret,
+			webhookSecret: GHES_WEBHOOK_SECRET,
 			privateKey: "myprivatekey",
 			installationId: 10
 		};
@@ -57,12 +73,7 @@ describe("webhook-receiver-post", () => {
 		res = {
 			sendStatus: jest.fn()
 		};
-		req = {
-			headers: {},
-			params: {
-				uuid: "97da6b0e-ec61-11ec-8ea0-0242ac120007"
-			}
-		};
+		req = createGHESReqForEvent("push", "", NON_EXIST_GHES_UUID);
 
 		await WebhookReceiverPost(req, res);
 		expect(res.sendStatus).toBeCalledWith(400);
@@ -75,15 +86,8 @@ describe("webhook-receiver-post", () => {
 				send: jest.fn()
 			})
 		};
-		req = {
-			headers: {
-				"x-hub-signature-256": "signature123"
-			},
-			params: {
-				uuid
-			},
-			body: {}
-		};
+		req = createGHESReqForEvent("push", "", EXIST_GHES_UUID);
+		req.headers["x-hub-signature-256"] = "invalid-signature";
 
 		await WebhookReceiverPost(req, res);
 		expect(res.status).toBeCalledWith(400);
@@ -97,14 +101,7 @@ describe("webhook-receiver-post", () => {
 				send: jest.fn()
 			})
 		};
-		req = {
-			headers: {
-				"x-hub-signature-256": "signature123"
-			},
-			params: {
-			},
-			body: {}
-		};
+		req = createReqWithInvalidSignature("push", undefined);
 
 		await WebhookReceiverPost(req, res);
 		expect(res.status).toBeCalledWith(400);
@@ -112,8 +109,35 @@ describe("webhook-receiver-post", () => {
 
 	});
 
+	describe("Pulling cloud or GHES app config", ()=>{
+		it("should pull cloud gitHubAppConfig with undefined UUID", async () => {
+			req = createCloudReqForEvent("push");
+			const spy = jest.fn();
+			jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
+			await WebhookReceiverPost(req, res);
+			expect(GithubWebhookMiddleware).toBeCalledWith(pushWebhookHandler);
+			expect(spy).toBeCalledWith(expect.objectContaining({
+				id: "100",
+				name: "push",
+				gitHubAppConfig: gitHubAppConfigForCloud()
+			}));
+		});
+		it("should pull GHES gitHubAppConfig with valid GHES UUID", async () => {
+			req = createGHESReqForEvent("push", "", EXIST_GHES_UUID);
+			const spy = jest.fn();
+			jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
+			await WebhookReceiverPost(req, res);
+			expect(GithubWebhookMiddleware).toBeCalledWith(pushWebhookHandler);
+			expect(spy).toBeCalledWith(expect.objectContaining({
+				id: "100",
+				name: "push",
+				gitHubAppConfig: gitHubAppConfigForES()
+			}));
+		});
+	});
+
 	it("should call push handler", async () => {
-		req = createReqForEvent("push");
+		req = createGHESReqForEvent("push", "", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -121,13 +145,13 @@ describe("webhook-receiver-post", () => {
 		expect(spy).toBeCalledWith(expect.objectContaining({
 			id: "100",
 			name: "push",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 
 	it("should call issue handler", async () => {
-		req = createReqForEvent("issues", "opened");
+		req = createGHESReqForEvent("issues", "opened", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -136,12 +160,12 @@ describe("webhook-receiver-post", () => {
 			id: "100",
 			name: "issues",
 			action: "opened",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 	it("should call pull handler", async () => {
-		req = createReqForEvent("pull_request", "opened");
+		req = createGHESReqForEvent("pull_request", "opened", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -150,12 +174,12 @@ describe("webhook-receiver-post", () => {
 			id: "100",
 			name: "pull_request",
 			action: "opened",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 	it("should call pull request review handler", async () => {
-		req = createReqForEvent("pull_request_review");
+		req = createGHESReqForEvent("pull_request_review", "", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -163,12 +187,12 @@ describe("webhook-receiver-post", () => {
 		expect(spy).toBeCalledWith(expect.objectContaining({
 			id: "100",
 			name: "pull_request_review",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 	it("should call create branch handler", async () => {
-		req = createReqForEvent("create");
+		req = createGHESReqForEvent("create", "", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -176,12 +200,12 @@ describe("webhook-receiver-post", () => {
 		expect(spy).toBeCalledWith(expect.objectContaining({
 			id: "100",
 			name: "create",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 	it("should call delete branch handler", async () => {
-		req = createReqForEvent("delete");
+		req = createGHESReqForEvent("delete", "", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -189,12 +213,12 @@ describe("webhook-receiver-post", () => {
 		expect(spy).toBeCalledWith(expect.objectContaining({
 			id: "100",
 			name: "delete",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 	it("should call delete repository handler", async () => {
-		req = createReqForEvent("repository", "deleted");
+		req = createGHESReqForEvent("repository", "deleted", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -203,12 +227,12 @@ describe("webhook-receiver-post", () => {
 			id: "100",
 			name: "repository",
 			action: "deleted",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 	it("should call workflow handler", async () => {
-		req = createReqForEvent("workflow_run");
+		req = createGHESReqForEvent("workflow_run", "", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -216,12 +240,12 @@ describe("webhook-receiver-post", () => {
 		expect(spy).toBeCalledWith(expect.objectContaining({
 			id: "100",
 			name: "workflow_run",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 	it("should call deployment handler", async () => {
-		req = createReqForEvent("deployment_status");
+		req = createGHESReqForEvent("deployment_status", "", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -229,12 +253,12 @@ describe("webhook-receiver-post", () => {
 		expect(spy).toBeCalledWith(expect.objectContaining({
 			id: "100",
 			name: "deployment_status",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 	it("should call code scanning handler", async () => {
-		req = createReqForEvent("code_scanning_alert");
+		req = createGHESReqForEvent("code_scanning_alert", "", EXIST_GHES_UUID);
 		const spy = jest.fn();
 		jest.mocked(GithubWebhookMiddleware).mockImplementation(() => spy);
 		await WebhookReceiverPost(req, res);
@@ -242,18 +266,52 @@ describe("webhook-receiver-post", () => {
 		expect(spy).toBeCalledWith(expect.objectContaining({
 			id: "100",
 			name: "code_scanning_alert",
-			gitHubAppConfig: gitHubAppConfig()
+			gitHubAppConfig: gitHubAppConfigForES()
 		}));
 	});
 
 });
 
-const createReqForEvent = (event: string, action?: string) => {
+const createReqWithInvalidSignature = (event: string, uuid?: string) => {
+	const body = {};
+
+	const req = {
+		headers: {
+			"x-hub-signature-256": "invalid-signature",
+			"x-github-event": event,
+			"x-github-delivery": "100"
+		},
+		params: {
+			uuid
+		},
+		body
+	};
+	return req;
+};
+
+const createCloudReqForEvent = (event: string, action?: string) => {
 	const body = action ? { action } : {};
 
 	const req = {
 		headers: {
-			"x-hub-signature-256": createHash(JSON.stringify(body), webhookSeret),
+			"x-hub-signature-256": createHash(JSON.stringify(body), CLOUD_WEBHOOK_SECRET),
+			"x-github-event": event,
+			"x-github-delivery": "100"
+		},
+		params: {
+			uuid: undefined
+		},
+		body
+	};
+	return req;
+};
+
+const createGHESReqForEvent = (event: string, action?: string, uuid?: string) => {
+	const body = action ? { action } : {};
+
+	const req = {
+		headers: {
+			"x-hub-signature-256": createHash(JSON.stringify(body), GHES_WEBHOOK_SECRET),
 			"x-github-event": event,
 			"x-github-delivery": "100"
 		},
